Create AP Spells
These are all the spells of the sf aps for you to copy&paste. they are working, i tested them. 

Note: i also include the aa3 spells (for my own use) until they are officially implemented. aa3 it self contains dimensional anchor as, well which was also published in #16. be aware of that.